+++ feed/post.html
@@ -10,6 +10,8 @@
   <link href="https://fonts.googleapis.com/css2?family=Mulish:ital,wght@0,200..1000;1,200..1000&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="../dist/css/index.css">
   <script defer src="../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
+  <script defer src="../js/post.js" type="module"></script>
+  <script defer src="../js/api/single-post-fetch.js" type="module"></script>
 </head>
 <body class="bg-login-body-green-white">
   <header>
@@ -46,10 +48,6 @@
         </div>
       </div>
     <div id="specific-post-content" class="mt-5">
-      <h1 class="text-frontpage font-fm-mulish border-bottom border-warning w-75 m-auto fs-1-5-rem pt-3 pb-2">Post title</h1>
-      <p class="text-frontpage font-fm-mulish fs-0-75rem-lg-1rem pt-3">Post-content</p>
-    </div>
-    <div class="feed-container row justify-content-center mt-3 mx-0">
     </div>
     </div>
   </div>
